backend/requirements.txt
fastapi
uvicorn
azure-storage-blob
pyodbc
pydantic
python-multipart
moviepy
imageio
imageio-ffmpeg
numpy


